refactor: move helper out of urls and into subject phase picker
